requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
google-api-python-client
python-dotenv
requests
plotly
isodate